moved samurai shift to enemy class parent
